policies Observes

+++ app/Http/Controllers/Admin/PostController.php
@@ -8,7 +8,7 @@ use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
-use App\Http\Requests\storePostRequest;
+use App\Http\Requests\PostRequest;
 class PostController extends Controller
 {
     /**
@@ -32,7 +32,7 @@ class PostController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(storePostRequest $request)
+    public function store(PostRequest $request)
     {
        
        $post = Post::create($request->all());
@@ -63,6 +63,9 @@ class PostController extends Controller
      */
     public function edit(Post $post )
         {
+
+
+            $this->authorize('author' , $post);
             $categories = Category::pluck('name', 'id');
             $tags = Tag::all();           
             return view('admin.posts.edit', compact('post','categories','tags'));
@@ -71,18 +74,37 @@ class PostController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Post $posts)
+    public function update(PostRequest $request, Post $post)
     {
-        //
+        $this->authorize('author' , $post);
+        $post->update($request->all());
+
+        if ($request->file('file')) {
+             $url =  Storage::put('public/posts', $request->file('file'));
+           
+             if($post->image){
+               Storage::delete($post->image->url);
+               $post->image->update([
+                'url' => $url
+               ]);
+             }else
+             {
+                $post->image()->create([
+                    'url' => $url
+                ]);
+             }
+        }
+        return redirect()->route('admin.posts.edit', $post )->with('info', 'El post Se ha actualizado');
     }
 
-    /**
+    /** 
      * Remove the specified resource from storage.
      */
     public function destroy(Post $post)
     {
         
+        $this->authorize('author' , $post);
         $post->delete();
-        return redirect()->route('admin.posts.index');
+        return redirect()->route('admin.posts.index')->with('info', 'El post Se ha eliminado Correctamente');
     }
 }
